refactor(hooks/use-auth-provider): move auth options to hooks
